+++ src/main/java/com/akkin/gulbi/application/GulbiService.java
@@ -4,11 +4,11 @@ import com.akkin.common.exception.GulbiNotFoundException;
 import com.akkin.common.exception.GulbiNotOwnerException;
 import com.akkin.gulbi.domain.Gulbi;
 import com.akkin.gulbi.persistence.GulbiRepository;
-import com.akkin.gulbi.dto.GulbiCreateForm;
-import com.akkin.gulbi.dto.GulbiCreateResponse;
-import com.akkin.gulbi.dto.GulbiReadResponses;
-import com.akkin.gulbi.dto.GulbiUpdateForm;
-import com.akkin.gulbi.dto.GulbiUpdateResponse;
+import com.akkin.gulbi.dto.request.GulbiCreateForm;
+import com.akkin.gulbi.dto.response.GulbiCreateResponse;
+import com.akkin.gulbi.dto.response.GulbiReadResponses;
+import com.akkin.gulbi.dto.request.GulbiUpdateForm;
+import com.akkin.gulbi.dto.response.GulbiUpdateResponse;
 import com.akkin.member.Member;
 import com.akkin.member.MemberService;
 import java.util.List;
