+++ src/main/java/com/akkin/gulbi/dto/response/GulbiReadResponse.java
@@ -1,4 +1,4 @@
-package com.akkin.gulbi.dto;
+package com.akkin.gulbi.dto.response;
 
 import com.akkin.gulbi.domain.Category;
 import com.akkin.gulbi.domain.Gulbi;
@@ -6,12 +6,11 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-@Schema(description = "서버에 저장된 아낀 정보")
+@Schema(description = "사용자가 입력한 아낀 항목 정보")
 @Getter
 @NoArgsConstructor
-public class GulbiCreateResponse {
+public class GulbiReadResponse {
 
-    @Schema(description = "아낀 항목 식별자")
     private Long id;
 
     private Integer year;
@@ -22,7 +21,7 @@ public class GulbiCreateResponse {
 
     private Category category;
 
-    private String saveContent;
+    private String content;
 
     private String how;
 
@@ -30,13 +29,13 @@ public class GulbiCreateResponse {
 
     private Integer realCost;
 
-    public GulbiCreateResponse(Gulbi gulbi) {
+    public GulbiReadResponse(Gulbi gulbi) {
         this.id = gulbi.getId();
         this.year = gulbi.getSaveYear();
         this.month = gulbi.getSaveMonth();
         this.day = gulbi.getSaveDay();
         this.category = gulbi.getCategory();
-        this.saveContent = gulbi.getContent();
+        this.content = gulbi.getContent();
         this.how = gulbi.getHow();
         this.expectCost = gulbi.getExpectCost();
         this.realCost = gulbi.getRealCost();
