+++ src/main/java/com/akkin/gulbi/presentation/GulbiApiController.java
@@ -3,11 +3,11 @@ package com.akkin.gulbi.presentation;
 import com.akkin.auth.aop.AuthRequired;
 import com.akkin.auth.dto.AuthMember;
 import com.akkin.gulbi.application.GulbiService;
-import com.akkin.gulbi.dto.GulbiCreateForm;
-import com.akkin.gulbi.dto.GulbiCreateResponse;
-import com.akkin.gulbi.dto.GulbiReadResponses;
-import com.akkin.gulbi.dto.GulbiUpdateForm;
-import com.akkin.gulbi.dto.GulbiUpdateResponse;
+import com.akkin.gulbi.dto.request.GulbiCreateForm;
+import com.akkin.gulbi.dto.response.GulbiCreateResponse;
+import com.akkin.gulbi.dto.response.GulbiReadResponses;
+import com.akkin.gulbi.dto.request.GulbiUpdateForm;
+import com.akkin.gulbi.dto.response.GulbiUpdateResponse;
 import javax.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
